rename a file for clarity
