+++ backend/server/utils/cqcLocationUtils.js
